Consolidate playback pages
