+++ app/src/main/java/com/github/damontecres/stashapp/ui/components/LongClicker.kt
@@ -18,7 +18,7 @@ import com.github.damontecres.stashapp.navigation.FilterAndPosition
 import com.github.damontecres.stashapp.navigation.NavigationManager
 import com.github.damontecres.stashapp.playback.PlaybackMode
 import com.github.damontecres.stashapp.ui.pages.DialogParams
-import com.github.damontecres.stashapp.ui.pages.maxPlaylistSize
+import com.github.damontecres.stashapp.ui.pages.MAX_PLAYLIST_SIZE
 import com.github.damontecres.stashapp.util.resume_position
 import java.util.concurrent.atomic.AtomicInteger
 import kotlin.time.Duration.Companion.seconds
@@ -133,7 +133,7 @@ class DefaultLongClicker(
                 }
                 if ((item is SlimSceneData || item is MarkerData) &&
                     filterAndPosition != null &&
-                    filterAndPosition.position < maxPlaylistSize // TODO
+                    filterAndPosition.position < MAX_PLAYLIST_SIZE // TODO
                 ) {
                     add(
                         DialogItem(
